backend/requirements.txt
# Core backend
flask
flask-cors
python-dotenv

# LangChain core + integrations
langchain
langchain-community
langchain-google-genai
langchain-huggingface

# Vector store + embeddings
faiss-cpu
sentence-transformers

# PDF & text handling
pypdf

# PyTorch (CPU only, needed for sentence-transformers on Windows)
torch
torchvision
torchaudio

# Optional utilities
requests
tqdm
